First cut of supporting OpenSSL 3.0 compiling
Compiling on Ubuntu 22.04, which is OpenSSL 3.0 based, this now compiles
and runs some basic unit tests, including talking to the server.

The changes were made such that older builds on OpenSSL 1.1.x still work
as before.

Some functions no longer exist in OpenSSL 3 and some new functions are
in OpenSSL 3 and not in 1.  Thus neither is a superset of the other.

This is currently a compile-time behavior and not a run-time behavior as
we used compile time compatibility shims and macros.

Doing this at run time will be more complex due to the current structure
of the code.

The containers and scripts I committed into ci/docker-dev for Carbon
take care of making it easy to test build both variants on the same
machine.  (Or more variants if we need to - I just built the 20.04 and
22.04 variants for OpenSSL 1.1 and OpenSSL 3.0 build and test
processes but adding RedHat or CentOS or Mariner or Debian would not be
a significant challenge.)  All of that is not part of this PR but makes
building and testing the variant versions much easier.

PS - I am sure that now that it is working we could write some code to
make it work in both but it is somewhat complicated by the fact that it
would take specific shims for a number of functions and parameters that
would need to dynamically determine the code paths to use for each
version.  It is non-trivial but it does not seem to be impossible.
(Tricky due to mapping of methods across different versions of header
files - unclear how to do that yet)